refactor:utils move from backend
